[SITEMAPS] Refactor UrlGenerator to be not document specific, but to only accept a path and add a document specific decorator
