+++ pimcore/lib/Pimcore/Bundle/CoreBundle/Resources/config/sitemaps.yml
@@ -13,6 +13,10 @@ services:
     #
     # Services valid for all elements
 
+    # Generic URL generator accepting a path
+    Pimcore\Sitemap\UrlGeneratorInterface: '@Pimcore\Sitemap\UrlGenerator'
+    Pimcore\Sitemap\UrlGenerator: ~
+
     # Filters
     Pimcore\Sitemap\Element\Filter\PublishedFilter: ~
     Pimcore\Sitemap\Element\Filter\PropertiesFilter: ~
@@ -33,8 +37,8 @@ services:
     Pimcore\Sitemap\Document\Filter\SiteRootFilter: ~
 
     # Default document URL generator
-    Pimcore\Sitemap\Document\UrlGeneratorInterface: '@Pimcore\Sitemap\Document\UrlGenerator'
-    Pimcore\Sitemap\Document\UrlGenerator: ~
+    Pimcore\Sitemap\Document\DocumentUrlGeneratorInterface: '@Pimcore\Sitemap\Document\DocumentUrlGenerator'
+    Pimcore\Sitemap\Document\DocumentUrlGenerator: ~
 
     # Default document tree generator
     Pimcore\Sitemap\Document\DocumentTreeGenerator:
